pyproject.toml
[tool.poetry]
name = "hyperbolic-agentkit-sample"
version = "0.1.0"
description = "A sample app that uses Hyperbolic AgentKit"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "0.104.1"
starlette = "0.27.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"
langchain-anthropic = "^0.3.0"
moviepy = "1.0.3"
opencv-python = "^4.11.0.86"
pillow = ">=9.2.0,<11.0"
pydantic = "^2.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
